+++ cypress/integration/app.spec.ts
@@ -1,4 +1,8 @@
-describe("Basic e2e", function () {
+import { FormInput } from "../support/types"
+
+import { DisplayObjectTypes, ObjectTypes } from "constants/wizardObject"
+
+describe("Basic application flow", function () {
   beforeEach(() => {
     cy.task("resetDb")
   })
@@ -7,60 +11,167 @@ describe("Basic e2e", function () {
     cy.login()
   })
 
-  it("should create new folder, add Study form, add Analysis form, add DAC form, and publish folder", () => {
+  it("completes all steps and publishes new submission", () => {
     cy.login()
 
     cy.get("button", { timeout: 10000 }).contains("Create submission").click()
 
+    /*
+     * 1st step, Submission details
+     */
+
     // Add folder name & description, navigate to submissions
     cy.newSubmission()
 
-    // Skip-link
-    cy.clickFillForm("Study")
+    /*
+     * 2nd step, Study, DAC and policy
+     */
 
-    // Upload a Study xml file.
-    cy.get("div[role=button]")
-      .contains("Upload XML File")
-      .should("be.visible")
-      .then($el => $el.click())
-    cy.fixture("study_test.xml").then(fileContent => {
-      cy.get("[data-testid='xml-upload']").attachFile(
-        {
-          fileContent: fileContent.toString(),
-          fileName: "testFile.xml",
-          mimeType: "text/xml",
-        },
-        { force: true }
-      )
+    // Try to send invalid form
+    cy.formActions("Submit")
+
+    cy.get<FormInput[]>("input[data-testid='descriptor.studyTitle']").then($input => {
+      expect($input[0].validationMessage).to.contain("Please fill")
     })
-    // Cypress doesn't allow form validation status to update and therefore "send" button is disabled
-    cy.get("form").submit()
 
-    // Saved objects list should have newly added item from Study object
-    cy.get(".MuiListItem-container", { timeout: 10000 }).should("have.length", 1)
+    // Fill a Study form and submit object
+    cy.get("[data-testid='descriptor.studyTitle']").type("Test title")
+    cy.get("[data-testid='descriptor.studyTitle']").should("have.value", "Test title")
 
-    // Replace the modified XML file
-    cy.get("button[type=button]")
-      .contains("Replace")
-      .should("be.visible")
-      .then($el => $el.click())
-    cy.get(".MuiCardHeader-action").contains("Replace")
-    cy.fixture("study_test_modified.xml").then(fileContent => {
-      cy.get("[data-testid='xml-upload']").attachFile(
-        {
-          fileContent: fileContent.toString(),
-          fileName: "testFile_replace.xml",
-          mimeType: "text/xml",
-        },
-        { force: true }
-      )
+    cy.formActions("Clear form")
+
+    cy.get("[data-testid='descriptor.studyTitle']").type("New title")
+    cy.get("[data-testid='descriptor.studyTitle']").should("have.value", "New title")
+    cy.get("[data-testid='descriptor.studyType']").select("Metagenomics")
+    cy.get("[data-testid='descriptor.studyAbstract']").type("New abstract")
+
+    // Submit form
+    cy.formActions("Submit")
+    cy.get("[data-testid='study-objects-list']").find("li").should("have.length", 1)
+
+    // DAC form
+    cy.clickAddObject(DisplayObjectTypes.dac)
+
+    // Try to submit empty DAC form. This should be invalid
+    cy.get("[data-testid=title]").type("Test title")
+    cy.formActions("Submit")
+    cy.get("span").contains("must have at least 1 item")
+
+    cy.get("h5").contains("Contacts").parents().children("button").click()
+    cy.get("[data-testid='contacts.0.name']").type("Test contact name")
+    // Test invalid email address (form array, default)
+    cy.get("[data-testid='contacts.0.email']").type("email")
+    cy.get("p[id='contacts.0.email-helper-text']").contains("must match format")
+
+    cy.get("[data-testid='contacts.0.email']").type("@test.com")
+    cy.get("[data-testid='contacts.0.telephoneNumber']").type("123456789")
+    cy.get("[data-testid='contacts.0.organisation-inputField']").type("Test organization")
+    // Click outside from organisation autocomplete field to hide suggestions
+    cy.get("[data-testid='contacts.0.telephoneNumber']").click()
+    cy.get("[data-testid='contacts.0.mainContact']").check()
+
+    // Submit DAC form
+    cy.formActions("Submit")
+
+    // Test DAC form update
+    cy.get("[data-testid='dac-objects-list']").within(() => {
+      cy.get("a").contains("Test contact name").click()
     })
-    cy.get("form").submit()
-    cy.get(".MuiListItem-container", { timeout: 10000 }).should("have.length", 1)
-    cy.contains(".MuiAlert-message", "Object replaced")
 
-    // Fill an Analysis form and submit object
-    cy.clickFillForm("Analysis")
+    cy.get("[data-testid='contacts']").should("be.visible")
+    cy.get("input[data-testid='contacts.0.name']", { timeout: 1000 }).should("be.visible").click({ force: true })
+    cy.get("input[data-testid='contacts.0.name']").type(" edited")
+    cy.get("[data-testid='contacts.0.name']").should("have.value", "Test contact name edited")
+
+    cy.formActions("Update")
+
+    cy.get("[data-testid='dac-objects-list']").within(() => {
+      cy.get("a").contains("Test contact name").click()
+    })
+    cy.get("[data-testid='contacts.0.name']").should("have.value", "Test contact name edited")
+
+    // Fill Policy form
+    cy.clickAddObject(ObjectTypes.policy)
+    cy.get("[data-testid='title']").type("Test Policy title")
+    cy.get("select[data-testid='dacRef.accessionId']").select(1)
+
+    cy.get("select[data-testid='dacRef.accessionId']").should("contain", " - Main Contact:")
+
+    cy.get("select[data-testid='policy']").select("Policy Text")
+    cy.get("textarea[data-testid='policy.policyText']").type("Test policy text")
+
+    // Submit Policy form
+    cy.formActions("Submit")
+
+    /*
+     * 3rd step, Describe
+     */
+
+    cy.clickAccordionPanel("Datafolder")
+    cy.get("[data-testid='datafolder-details']", { timeout: 10000 }).should("be.visible")
+    cy.get("button[role=button]", { timeout: 10000 }).contains("Link datafolder").click()
+
+    /*
+     * 4th step, Datafolder
+     */
+    cy.clickAccordionPanel("Describe")
+
+    // Fill a Sample form
+    cy.clickAddObject(ObjectTypes.sample)
+    cy.get("[data-testid='title']").type("Test Sample title")
+    cy.get("[data-testid='sampleName.taxonId']").type("123")
+    // Submit Sample form
+    cy.formActions("Submit")
+
+    // Fill a Experiment form
+    cy.clickAddObject(ObjectTypes.experiment)
+    cy.get("[data-testid='title']").type("Test Experiment title")
+
+    cy.get("select[data-testid='platform']").select("AB 5500 Genetic Analyzer")
+    cy.get("select[data-testid='platform']").should("have.value", "AB 5500 Genetic Analyzer")
+
+    // Select studyAccessionId
+    cy.get("select[data-testid='studyRef.accessionId']").select(1)
+
+    cy.get("select[data-testid='studyRef.accessionId']").should("contain", " - Title:")
+
+    cy.get("textarea[data-testid='design.designDescription']").type("Design description")
+    cy.get("select[data-testid='design.sampleDescriptor']").select("Individual Sample")
+    cy.get("select[data-testid='design.sampleDescriptor']").should("have.value", "Individual Sample")
+
+    // Select sampleAccessionId
+    cy.get("select[data-testid='design.sampleDescriptor.accessionId']").select(1)
+
+    cy.get("select[data-testid='design.sampleDescriptor.accessionId']").should("contain", " - Title:")
+
+    cy.get("select[data-testid='design.libraryDescriptor.libraryStrategy']").select("AMPLICON")
+    cy.get("select[data-testid='design.libraryDescriptor.libraryStrategy']").should("have.value", "AMPLICON")
+    cy.get("select[data-testid='design.libraryDescriptor.librarySource']").select("GENOMIC")
+    cy.get("select[data-testid='design.libraryDescriptor.librarySource']").should("have.value", "GENOMIC")
+    cy.get("select[data-testid='design.libraryDescriptor.librarySelection']").select("CAGE")
+    cy.get("select[data-testid='design.libraryDescriptor.librarySelection']", { timeout: 10000 }).should(
+      "have.value",
+      "CAGE"
+    )
+
+    // Submit Experiment form
+    cy.formActions("Submit")
+
+    // Fill a Run form
+    cy.clickAddObject(ObjectTypes.run)
+    cy.get("[data-testid='title']").type("Test Run title")
+    cy.get("h5[data-testid='experimentRef']").parents().children("button").click()
+
+    // Select experimentAccessionId
+    cy.get("select[data-testid='experimentRef.0.accessionId']").select(1)
+
+    cy.get("select[data-testid='experimentRef.0.accessionId']").should("contain", " - Title:")
+
+    // Submit Run form
+    cy.formActions("Submit")
+
+    // Fill an Analysis form
+    cy.clickAddObject(ObjectTypes.analysis)
 
     cy.get("form").within(() => {
       cy.get("[data-testid='title']").type("Test title")
@@ -75,24 +186,25 @@ describe("Basic e2e", function () {
       )
 
       // Study
-      cy.get("[data-testid='studyRef.accessionId']").type("Study Test Accession Id")
+      // Select studyAccessionId
+      cy.get("select[data-testid='studyRef.accessionId']").select(1)
       cy.get("[data-testid='studyRef.refname']").type("Study Test Record Name")
       cy.get("[data-testid='studyRef.refcenter']").type("Study Test Namespace")
 
       // Experiment
-      cy.get("[data-testid='experimentRef.accessionId']").type("Experiment Test Accession Id")
+      cy.get("[data-testid='experimentRef.accessionId']").select(1)
       cy.get("[data-testid='experimentRef.refcenter']").type("Experiment Test Record Name")
       cy.get("[data-testid='experimentRef.refname']").type("Experiment Test Namespace")
 
       // Sample
       cy.get("h5").contains("Sample Reference").parents().children("button").click()
-      cy.get("[data-testid='sampleRef.0.accessionId']").type("Sample Test Accession Id")
+      cy.get("[data-testid='sampleRef.0.accessionId']").select(1)
       cy.get("[data-testid='sampleRef.0.identifiers.submitterId.namespace']").type("Sample Test Namespace")
       cy.get("[data-testid='sampleRef.0.identifiers.submitterId.value']").type("Sample Test Value")
 
       // Run
       cy.get("h5").contains("Run Reference").parents().children("button").click()
-      cy.get("[data-testid='runRef.0.accessionId']").type("Run Test Accession Id")
+      cy.get("[data-testid='runRef.0.accessionId']").select(1)
       cy.get("[data-testid='runRef.0.identifiers.submitterId.namespace']").type("Run Test Namespace")
       cy.get("[data-testid='runRef.0.identifiers.submitterId.value']").type("Run Test Value")
 
@@ -117,58 +229,50 @@ describe("Basic e2e", function () {
         "c34045c1a1db8d1b3fca8a692198466952daae07eaf6104b4c87ed3b55b6af1b"
       )
     })
-    // Submit form
+    // Submit Analysis form
     cy.formActions("Submit")
-    // Saved objects list should have newly added item from Analysis object
-    cy.get(".MuiListItem-container", { timeout: 10000 }).should("have.length", 1)
 
-    // DAC form
-    cy.clickFillForm("DAC")
+    // Fill a Dataset form
+    cy.clickAddObject(ObjectTypes.dataset)
 
-    // Try to submit empty DAC form. This should be invalid
-    cy.get("[data-testid=title]").type("test description")
+    cy.get("[data-testid='title']").type("Test Dataset title")
+    cy.get("[data-testid='description']").type("Dataset description")
+    cy.get("[data-testid='datasetType']").first().check()
+
+    // Select policyAccessionId
+    cy.get("select[data-testid='policyRef.accessionId']").select(1)
+    cy.get("select[data-testid='policyRef.accessionId']").should("contain", " - Title:")
+
+    // Select runAccessionId
+    cy.get("div").contains("Run Reference").parents().children("button").click()
+    cy.get("select[data-testid='runRef.0.accessionId']").select(1)
+    cy.get("select[data-testid='runRef.0.accessionId']").should("contain", " - Title:")
+
+    // Select analysisAccessionId
+    cy.get("div").contains("Analysis Reference").parents().children("button").click()
+    cy.get("select[data-testid='analysisRef.0.accessionId']").select(1)
+    cy.get("select[data-testid='analysisRef.0.accessionId']").should("contain", " - Title:")
+
+    // Submit Dataset form
     cy.formActions("Submit")
-    cy.get("span").contains("must have at least 1 item")
 
-    cy.get("h5").contains("Contacts").parents().children("button").click()
-    cy.get("[data-testid='contacts.0.name']").type("Test contact name")
-    // Test invalid email address (form array, default)
-    cy.get("[data-testid='contacts.0.email']").type("email")
-    cy.get("p[id='contacts.0.email-helper-text']").contains("must match format")
-
-    cy.get("[data-testid='contacts.0.email']").type("@test.com")
-    cy.get("[data-testid='contacts.0.telephoneNumber']").type("123456789")
-    cy.get("[data-testid='contacts.0.organisation-inputField']").type("Test organization")
-    // Click outside from organisation autocomplete field to hide suggestions
-    cy.get("[data-testid='contacts.0.telephoneNumber']").click()
-    cy.get("[data-testid='contacts.0.mainContact']").check()
-
-    // Submit form
-    cy.formActions("Submit")
-    // Saved objects list should have newly added item from DAC object
-    cy.get(".MuiListItem-container", { timeout: 10000 }).should("have.length", 1)
-
-    // Test DAC form update
-    cy.get("button[type=button]").contains("Edit").click()
-    cy.get("[data-testid='contacts']").should("be.visible")
-    cy.get("input[data-testid='contacts.0.name']", { timeout: 1000 }).should("be.visible").click({ force: true })
-    cy.get("input[data-testid='contacts.0.name']").type(" edited")
-    cy.get("[data-testid='contacts.0.name']").should("have.value", "Test contact name edited")
-
-    cy.formActions("Update")
-
-    cy.get("button[type=button]").contains("Edit").click()
-    cy.get("[data-testid='contacts.0.name']").should("have.value", "Test contact name edited")
-
-    // Navigate to summary
-    cy.get("button[type=button]").contains("Next").click()
+    /*
+     * 5th step, Identifier and publish
+     */
+    cy.clickAccordionPanel("publish")
+    cy.get("button[role=button]", { timeout: 10000 }).contains("Publish").click()
 
     // Check the amount of submitted objects in each object type
     cy.get("h6").contains("Study").parents("div").children().eq(1).should("have.text", 1)
+    cy.get("h6").contains("DAC").parents("div").children().eq(1).should("have.text", 1)
+    cy.get("h6").contains("Policy").parents("div").children().eq(1).should("have.text", 1)
+    cy.get("h6").contains("Sample").parents("div").children().eq(1).should("have.text", 1)
+    cy.get("h6").contains("Experiment").parents("div").children().eq(1).should("have.text", 1)
+    cy.get("h6").contains("Run").parents("div").children().eq(1).should("have.text", 1)
     cy.get("h6").contains("Analysis").parents("div").children().eq(1).should("have.text", 1)
+    cy.get("h6").contains("Dataset").parents("div").children().eq(1).should("have.text", 1)
 
-    // Navigate to publish
-    cy.get("button[type=button]").contains("Publish").click()
+    cy.get("button[data-testid='summary-publish']").contains("Publish").click()
     cy.get("[data-testid='alert-dialog-content']").should("have.text", "Objects in this folder will be published.")
     cy.get('[data-testid="confirm-publish-folder"]').contains("Publish").click()
   })